Move update logic to update.ts and update tests
